requirements.txt
ffmpeg-python
gifpy